Move SocketProvider under provider directory
